+++ builder/builders/compilers/index.ts
@@ -1,22 +1,24 @@
 import { IBuilderContext } from './../../utils';
-import { ExportType } from './../utils';
-import { CompileCSS } from './css';
-import { CompileTWIG } from './twig';
-import { CompileTS } from './typescript';
-
-export type Compile = (file: string, output: string, context: IBuilderContext) => Promise<{ path: string, type: ExportType }>;
+import { Compiler } from './compiler';
+import { CSSCompiler } from './css';
+import { TWIGCompiler } from './twig';
+import { TSCompiler } from './typescript';
 
 const CompilerMapping : {
-  [key: string]: Compile
+  [key: string]: Compiler
 } = {
-  ts: CompileTS,
-  css: CompileCSS,
-  twig: CompileTWIG
+  ts: new TSCompiler(),
+  css: new CSSCompiler(),
+  twig: new TWIGCompiler()
 }
 
 // TODO: dynamically generate the union type based on mapping keys
 export type CompielableType = 'ts' | 'css' | 'twig';
 
 export async function Compile(type: CompielableType, file: string, exportPath: string, context: IBuilderContext) {
-  return await CompilerMapping[type](file, exportPath, context);
+  return CompilerMapping[type].compile(file, exportPath, context);
+}
+
+export function getContextFiles(type: CompielableType, file: string, exportPath: string, context: IBuilderContext) {
+  return CompilerMapping[type].getContextFiles(file, exportPath, context);
 }