+++ cli/watch.ts
@@ -1,10 +1,11 @@
 import { Command } from 'commander';
 import fs from 'fs';
+import chalk from 'chalk';
 
 import { Builder } from './../builder';
 import { build, generateConfig, IMainCommanderOptions } from './utils';
 
-const chalk = require('chalk');
+// const chalk = require('chalk');
 
 const log = console.log;
 
@@ -13,12 +14,13 @@ export function make(program: Command) {
 
   heat
     .action(async () => {
-      log('Watching...');
-
       const config = generateConfig(program.opts() as IMainCommanderOptions);
       const builder = new Builder(config);
 
-      config.files.forEach(file => {
+      log('Watching files');
+      builder.contextFiles.forEach(v => log('  ' + chalk.underline.blue(v)))
+
+      builder.contextFiles.forEach(file => {
         fs.watchFile(file, () => build(builder));
       });
       build(builder);